+++ src/main/java/com/finalproject/manitoone/notification/constants/NotiType.java
@@ -1,4 +1,4 @@
-package com.finalproject.manitoone.notifications.constants;
+package com.finalproject.manitoone.notification.constants;
 
 public enum NotiType {
